+++ projects/packages/forms/src/class-wpcom-rest-api-v2-endpoint-forms.php
@@ -9,12 +9,16 @@
 namespace Automattic\Jetpack\Forms;
 
 use Automattic\Jetpack\Connection\Manager;
+use Automattic\Jetpack\Forms\ContactForm\Contact_Form_File_Handler;
 use Automattic\Jetpack\Forms\ContactForm\Contact_Form_Plugin;
 use WP_Error;
 use WP_REST_Controller;
 use WP_REST_Request;
 use WP_REST_Response;
 use WP_REST_Server;
+
+// Require the file handler class for file endpoint functionality
+require_once __DIR__ . '/contact-form/class-contact-form-file-handler.php';
 
 /**
  * Handles the REST routes for Form Responses, aka Feedback.
@@ -78,6 +82,27 @@ class WPCOM_REST_API_V2_Endpoint_Forms extends WP_REST_Controller {
 				'methods'             => WP_REST_Server::CREATABLE,
 				'callback'            => array( $this, 'bulk_actions' ),
 				'permission_callback' => array( $this, 'get_responses_permission_check' ),
+			)
+		);
+
+		// Register the file endpoint route
+		register_rest_route(
+			'jetpack-forms/v1',
+			'files',
+			array(
+				'methods'                 => WP_REST_Server::READABLE,
+				'callback'                => array( $this, 'get_file' ),
+				'permission_callback'     => array( $this, 'get_file_permissions_check' ),
+				'args'                    => array(
+					'file_id' => array(
+						'required'          => true,
+						'validate_callback' => function ( $param ) {
+							return ! empty( $param );
+						},
+						'sanitize_callback' => 'sanitize_text_field',
+					),
+				),
+				'requires_authentication' => true,
 			)
 		);
 	}
@@ -486,6 +511,116 @@ class WPCOM_REST_API_V2_Endpoint_Forms extends WP_REST_Controller {
 	 */
 	private function error_response( $message, $code ) {
 		return new WP_REST_Response( array( 'error' => $message ), $code );
+	}
+
+	/**
+	 * Checks if the current user has permission to view files.
+	 *
+	 * @param \WP_REST_Request $request The current request object.
+	 * @return true|\WP_Error True if the user has permission, WP_Error otherwise.
+	 */
+	public function get_file_permissions_check( $request ) {
+		// Verify the user is logged in with appropriate capabilities
+		if ( ! current_user_can( 'edit_pages' ) ) {
+			return new WP_Error(
+				'rest_forbidden',
+				esc_html__( 'You must be logged in with appropriate permissions to view this file.', 'jetpack-forms' ),
+				array( 'status' => 403 )
+			);
+		}
+
+		// Get the file ID from the request and its hash
+		$file_id      = $request->get_param( 'file_id' );
+		$file_id_hash = $request->get_param( 'file_id_hash' );
+
+		// If no hash was provided, generate it from the file_id
+		if ( empty( $file_id_hash ) ) {
+			$file_id_hash = md5( $file_id );
+		}
+
+		// Verify the file-specific nonce using the hash
+		$file_nonce = $request->get_param( 'file_nonce' );
+		if ( ! $file_nonce || ! wp_verify_nonce( $file_nonce, 'jetpack_forms_view_file_' . $file_id_hash ) ) {
+			return new WP_Error(
+				'rest_forbidden',
+				esc_html__( 'Invalid or missing file access token.', 'jetpack-forms' ),
+				array( 'status' => 403 )
+			);
+		}
+
+		return true;
+	}
+
+	/**
+	 * Retrieves a file using the file_id and serves it to the client.
+	 *
+	 * @param \WP_REST_Request $request The current request object.
+	 * @return \WP_REST_Response|\WP_Error Response object or error.
+	 */
+	public function get_file( $request ) {
+		$file_id = $request->get_param( 'file_id' );
+
+		// Initialize the file handler
+		$file_handler = new Contact_Form_File_Handler();
+
+		// Get the full file path
+		$file_path = $file_handler->get_file_path( $file_id );
+
+		if ( empty( $file_path ) || ! file_exists( $file_path ) ) {
+			// Log the failure for debugging using WordPress logging
+			if ( defined( 'WP_DEBUG' ) && WP_DEBUG ) {
+				// Use apply_filters instead of direct error_log for debugging
+				do_action( 'jetpack_forms_debug_message', sprintf( 'Jetpack Forms: File not found. ID: %s, Path: %s', $file_id, $file_path ) );
+			}
+
+			return new WP_Error(
+				'file_not_found',
+				esc_html__( 'The requested file does not exist.', 'jetpack-forms' ),
+				array( 'status' => 404 )
+			);
+		}
+
+		// Get the file mime type
+		$mime_type = $this->get_file_mime_type( $file_path );
+
+		// Use WP_Filesystem to read and output the file instead of readfile()
+		require_once ABSPATH . 'wp-admin/includes/file.php';
+		WP_Filesystem();
+		global $wp_filesystem;
+
+		nocache_headers();
+		header( 'X-Robots-Tag: noindex', true );
+		header( 'Content-Type: ' . $mime_type );
+		header( 'Content-Description: File Transfer' );
+		header( 'Content-Disposition: inline; filename="' . wp_basename( $file_path ) . '"' );
+		header( 'Content-Transfer-Encoding: binary' );
+		header( 'Content-Length: ' . $wp_filesystem->size( $file_path ) );
+		header( 'Cache-Control: must-revalidate, post-check=0, pre-check=0' );
+
+		// Clear any previous output buffers
+		while ( ob_get_level() ) {
+			ob_end_clean();
+		}
+
+		if ( $wp_filesystem->exists( $file_path ) ) {
+			echo $wp_filesystem->get_contents( $file_path ); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped -- File contents should not be escaped
+		}
+		exit;
+	}
+
+	/**
+	 * Get the mime type of a file.
+	 *
+	 * @param string $file_path Path to the file.
+	 * @return string The mime type.
+	 */
+	protected function get_file_mime_type( $file_path ) {
+		$mime_type = mime_content_type( $file_path );
+		if ( false === $mime_type ) {
+			// Fallback to a generic mime type
+			$mime_type = 'application/octet-stream';
+		}
+		return $mime_type;
 	}
 }
